% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{buildSet}
\alias{buildSet}
\title{Read Dataset}
\usage{
buildSet(fpath = NULL, file_list = NULL, sample_list = NULL,
  read_col = "TPM", concat = TRUE)
}
\arguments{
\item{fpath}{Path to directory of files to read from. If not set then file_list must be set.}

\item{file_list}{Character vector of filepaths of files to be read from.}

\item{sample_list}{Character vector, sample names to associate with each sample in file_list}

\item{read_col}{Column in the provided files to read from.}

\item{concat}{Whether to read all files at once and then merge (faster but more memory inefficient) or to merge file by file.}
}
\value{
Single dataframes where each col corresponds to the readings from a single sample.
}
\description{
Reads a dataset either from a list of files or from a single directory.
}
\details{
Two options are available. The first option (default) reads all samples and merges them afterward while
the second option creates the dataset file by file which is slower but less memory intensive.
}
